fetch/api/request/request-navigation-type.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<title>Request navigation type attributes</title>
<meta name="help" href="https://fetch.spec.whatwg.org/#request">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
</head>
<body>
<script>
test(() => {
  assert_false(new Request('/').isReloadNavigation, 'isReloadNavigation');
}, 'default values for navigation type attributes');
</script>
</body>
</html>
